+++ OjtPageHandler.inc.php
@@ -18,8 +18,13 @@ class OjtPageHandler extends Handler
         parent::__construct();
 
         $this->addRoleAssignment(
-            [ROLE_ID_SITE_ADMIN, ROLE_ID_MANAGER],
+            ROLE_ID_SITE_ADMIN,
             ['index', 'getInstalledPlugin', 'updatePanel', 'settings', 'saveSettings', 'downloadLog', 'reportBug', 'submitBug', 'checkUpdate', 'getPluginGalleryList', 'getExclusivePlugins', 'save', 'installPlugin', 'uninstallPlugin', 'checkPluginInstalled', 'toggleInstalledPlugin', 'resetSetting', 'support'],
+        );
+
+        $this->addRoleAssignment(
+            ROLE_ID_MANAGER,
+            ['index', 'getInstalledPlugin', 'updatePanel', 'settings', 'saveSettings', 'downloadLog', 'reportBug', 'submitBug', 'checkUpdate', 'getPluginGalleryList', 'getExclusivePlugins', 'save', 'installPlugin', 'checkPluginInstalled', 'toggleInstalledPlugin', 'resetSetting', 'support'],
         );
 
         $this->ojtPlugin = OjtPlugin::get();
@@ -113,6 +118,7 @@ class OjtPageHandler extends Handler
         HookRegistry::call('OjtPageHandler::index', array(&$ojtPlugin));
 
         $templateMgr->assign('ojtPlugin', $ojtPlugin);
+        $templateMgr->assign('canDeletePlugins', $plugin->isCurrentUserSiteAdmin());
         $templateMgr->assign('journal', $this->contextId ? $request->getContext() : $request->getSite());
         $templateMgr->assign('pluginGalleryHtml', $templateMgr->fetch($this->ojtPlugin->getTemplateResource('plugingallery.tpl')));
         $templateMgr->assign('pluginInstalledHtml', $templateMgr->fetch($this->ojtPlugin->getTemplateResource('plugininstalled.tpl')));
@@ -193,7 +199,7 @@ class OjtPageHandler extends Handler
     public function submitBug($args, $request)
     {
         try {
-            $url = 'https://sp.openjournaltheme.com/api/v1/report';
+            $url = 'http://127.0.0.1:8000/api/v1/report';
 
             $params = $request->getUserVars();
             $files = $this->reArrayFiles($_FILES['pictures']);
@@ -385,10 +391,9 @@ class OjtPageHandler extends Handler
         HookRegistry::call('OjtPageHandler::installed::plugins', array($this, &$plugins));
 
         // Ensure canDelete and isAuthorized are set
+        $isSiteAdmin = $this->ojtPlugin->isCurrentUserSiteAdmin();
         foreach($plugins as &$plugin) {
-            if(!isset($plugin['canDelete'])) {
-                $plugin['canDelete'] = true;
-            }
+            $plugin['canDelete'] = $isSiteAdmin && ($plugin['canDelete'] ?? true);
             if(!isset($plugin['isAuthorized'])) {
                 $plugin['isAuthorized'] = true;
             }
@@ -727,6 +732,13 @@ class OjtPageHandler extends Handler
     public function uninstallPlugin($args, $request)
     {
         $plugin = $this->ojtPlugin;
+
+        if (!$plugin->isCurrentUserSiteAdmin()) {
+            $json['error'] = 1;
+            $json['msg'] = 'Only site administrators can uninstall plugins';
+            showJson($json);
+            return;
+        }
 
         $removePlugin = json_decode($request->getUserVar('plugin'));